@@ -1,80 +1,47 @@
 {
-  "name": "@libary/source",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {},
+  "name": "shop-eazy-01",
   "private": true,
-  "devDependencies": {
-    "@eslint/js": "^9.8.0",
-    "@nrwl/react": "^19.8.4",
-    "@nx/eslint": "19.8.4",
-    "@nx/eslint-plugin": "19.8.4",
-    "@nx/jest": "19.8.4",
-    "@nx/js": "20.8.0",
-    "@nx/react": "19.8.4",
-    "@nx/storybook": "19.8.4",
-    "@nx/vite": "19.8.4",
-    "@nx/web": "19.8.4",
-    "@storybook/addon-essentials": "^8.2.8",
-    "@storybook/addon-interactions": "^8.2.8",
-    "@storybook/core-server": "^8.2.8",
-    "@storybook/jest": "^0.2.3",
-    "@storybook/react-vite": "^8.2.8",
-    "@storybook/test-runner": "^0.13.0",
-    "@storybook/testing-library": "^0.2.2",
-    "@swc-node/register": "~1.9.1",
-    "@swc/cli": "~0.3.12",
-    "@swc/core": "~1.5.7",
-    "@swc/helpers": "~0.5.11",
-    "@testing-library/react": "15.0.6",
-    "@types/jest": "^29.5.12",
-    "@types/node": "18.16.9",
-    "@types/react": "18.3.1",
-    "@types/react-dom": "18.3.0",
-    "@vitejs/plugin-react": "^4.2.0",
-    "@vitest/ui": "^1.3.1",
-    "babel-jest": "^29.7.0",
-    "eslint": "^9.8.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "2.30.0",
-    "eslint-plugin-jsx-a11y": "6.7.1",
-    "eslint-plugin-react": "7.35.0",
-    "eslint-plugin-react-hooks": "4.6.0",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "jsdom": "~22.1.0",
-    "nx": "20.8.0",
-    "prettier": "^2.6.2",
-    "storybook": "^8.2.8",
-    "ts-jest": "^29.1.0",
-    "ts-node": "10.9.1",
-    "tslib": "^2.3.0",
-    "typescript": "~5.7.2",
-    "typescript-eslint": "^8.0.0",
-    "vite": "^5.0.0",
-    "vitest": "^1.3.1"
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc -b && vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
   },
-  "workspaces": [
-    "packages/*"
-  ],
   "dependencies": {
     "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.0",
-    "@fortawesome/fontawesome-svg-core": "^6.7.2",
-    "@fortawesome/free-brands-svg-icons": "^6.7.2",
-    "@fortawesome/free-solid-svg-icons": "^6.7.2",
-    "@fortawesome/react-fontawesome": "^0.2.2",
-    "@mui/material": "^7.0.2",
-    "bootstrap": "^5.3.5",
-    "install": "^0.13.0",
+    "@emotion/styled": "^11.14.1",
+    "@fortawesome/fontawesome-free": "^7.0.1",
+    "@fortawesome/free-brands-svg-icons": "^7.0.1",
+    "@fortawesome/free-regular-svg-icons": "^7.0.1",
+    "@fortawesome/free-solid-svg-icons": "^7.0.1",
+    "@fortawesome/react-fontawesome": "^3.0.2",
+    "@mui/material": "^7.3.2",
+    "@reduxjs/toolkit": "^2.9.0",
+    "axios": "^1.12.2",
+    "bootstrap": "^5.3.8",
+    "bootstrap-icons": "^1.13.1",
     "jquery": "^3.7.1",
-    "npm": "^11.4.1",
-    "owl.carousel": "^2.3.4",
-    "react": "18.3.1",
-    "react-bootstrap": "^2.10.9",
-    "react-dom": "18.3.1",
-    "react-owl-carousel2": "^0.3.0",
-    "react-router-dom": "^7.6.1",
-    "swiper": "^11.2.8"
+    "json-server": "^1.0.0-beta.3",
+    "react": "^19.1.1",
+    "react-bootstrap": "^2.10.10",
+    "react-dom": "^19.1.1",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^7.9.1",
+    "swiper": "^12.0.2"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.36.0",
+    "@types/react": "^19.1.13",
+    "@types/react-dom": "^19.1.9",
+    "@vitejs/plugin-react": "^5.0.3",
+    "eslint": "^9.36.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.20",
+    "globals": "^16.4.0",
+    "typescript": "~5.8.3",
+    "typescript-eslint": "^8.44.0",
+    "vite": "^7.1.7"
   }
 }
